exception_details: split into exception_details_invalid
